requirements.txt
flask
torch
torchvision
opencv-python
pillow
gunicorn